refactor to stream window
